# apps/api/main.py
import asyncio
import logging.config
import uuid
from contextlib import asynccontextmanager, suppress

from integrations.sentry import init_sentry
from auth import CSRF_COOKIE_NAME, ENABLE_CSRF
from billing.routes import billing_router
from config import settings
from database import init_db
from fastapi import Depends, FastAPI, HTTPException, Request, status
from fastapi.middleware.cors import CORSMiddleware
from fastapi.responses import JSONResponse, Response
from log_config import LOGGING_CONFIG, clear_log_context, reset_request_id, set_request_id
from middleware.body_limit import BodySizeLimitMiddleware
from parliament.model_registry import get_default_model, list_enabled_models
from promotions.routes import promotions_router
from ratelimit import ensure_rate_limiter_ready
from routes.admin import (
    admin_logs,
    admin_ops_summary,
    admin_router,
    admin_users,
    update_ratings_endpoint,
)
from routes.api_keys import api_keys_router
from routes.auth import (
    AuthRequest,
    auth_router,
    get_me,
    google_callback,
    google_login,
    login_user,
    logout_user,
    register_user,
)
from routes.debates import (
    DebateUpdate,
    create_debate,
    debates_router,
    export_debate_report,
    export_scores_csv,
    get_debate,
    get_debate_events,
    get_debate_judges,
    get_debate_report,
    get_leaderboard,
    get_leaderboard_persona,
    list_debates,
    start_debate_run,
    stream_events,
    update_debate,
)
from routes.models import models_router
from routes.ops import router as ops_router
from routes.stats import (
    get_debate_summary,
    get_hall_of_fame_stats,
    get_model_detail,
    get_model_leaderboard_stats,
    get_rate_limit_snapshot,
    get_system_health,
    stats_router,
)
from routes.teams import (
    TeamCreate,
    TeamMemberCreate,
    add_team_member,
    create_team,
    list_team_members,
    list_teams,
    teams_router,
)
from schemas import DebateCreate
from sse_backend import get_sse_backend
from starlette.middleware.base import BaseHTTPMiddleware

root_level = settings.LOG_LEVEL.upper()
# MyPy sees LOGGING_CONFIG as Dict[str, Any] but inferred loosely.
# We can just ignore the type error for this dynamic configuration or cast it.
LOGGING_CONFIG["loggers"][""]["level"] = root_level  # type: ignore
for logger_name in ("uvicorn", "uvicorn.error", "uvicorn.access", "apps"):
    if logger_name in LOGGING_CONFIG["loggers"]:
        LOGGING_CONFIG["loggers"][logger_name]["level"] = root_level  # type: ignore
logging.config.dictConfig(LOGGING_CONFIG)
logger = logging.getLogger(__name__)
TEST_FAST_APP = settings.ENV == "test"

# Safety guard: prevent test mode in non-local environments
if TEST_FAST_APP and not settings.IS_LOCAL_ENV:
    raise RuntimeError(
        "FATAL: ENV='test' is set in a non-local environment. "
        "This would disable CSRF, security headers, and other protections. "
        "Refusing to start. Set ENV to 'production' or 'staging'."
    )

if TEST_FAST_APP:

    try:
        from database import reset_engine as _reset_engine
    except ImportError:
        logger.warning("TEST_FAST_APP enabled but reset_engine is not available")
    else:
        try:
            _reset_engine()
        except Exception as exc:
            logger.error("Failed to reset DB engine in TEST_FAST_APP mode", exc_info=exc)
            raise

init_sentry()

SAFE_METHODS = {"GET", "HEAD", "OPTIONS", "TRACE"}
ENABLE_SEC_HEADERS = settings.ENABLE_SEC_HEADERS


async def csrf_protect(request: Request) -> None:
    """Optional double-submit CSRF guard for cookie auth.
    
    CSRF protection is skipped when:
    - Request uses a safe HTTP method
    - ENABLE_CSRF is False
    - Request is to auth endpoints (login/register)
    - Request uses Bearer token auth (Authorization header)
    """
    if request.method in SAFE_METHODS or not ENABLE_CSRF:
        return
    if request.url.path in {"/auth/login", "/auth/register"}:
        return
    # Skip CSRF for Bearer token auth - CSRF attacks only affect cookie auth
    auth_header = request.headers.get("Authorization", "")
    if auth_header.startswith("Bearer "):
        return
    csrf_cookie = request.cookies.get(CSRF_COOKIE_NAME)
    csrf_header = request.headers.get("x-csrf-token") or request.headers.get("X-CSRF-Token")
    if not csrf_cookie or not csrf_header or csrf_cookie != csrf_header:
        raise HTTPException(
            status_code=status.HTTP_403_FORBIDDEN,
            detail="CSRF token missing or invalid",
        )


@asynccontextmanager
async def lifespan(app: FastAPI):
    if TEST_FAST_APP:
        # Ensure sse_backend is available in state for dependencies even in test mode
        from sse_backend import get_sse_backend as get_global_sse
        app.state.sse_backend = get_global_sse()
        yield
        return
    init_db()
    # Verify that critical tables exist; retry briefly if DB is warming up
    try:
        from sqlalchemy import create_engine, inspect
        engine = create_engine(settings.DATABASE_URL)
        critical_tables = {
            "user",
            "team_member",
            "usage_quota",
            "usage_counter",
            "api_keys",
            "audit_log",
            "debate",
        }
        
        # Bounded retry for database warm-up
        max_retries = 15
        retry_delay = 2
        missing = []
        
        for attempt in range(max_retries):
            try:
                inspector = inspect(engine)
                existing_tables = set(inspector.get_table_names())
                missing = [tbl for tbl in critical_tables if tbl not in existing_tables]
                if not missing:
                    logger.info("Database connection and schema verified successfully.")
                    break
                else:
                    logger.warning("Attempt %d: Missing critical tables: %s (DB may be mid-migration)", attempt + 1, missing)
            except Exception as e:
                logger.warning("Attempt %d: Database connection failed (warming up?): %s", attempt + 1, e)
                
            if attempt < max_retries - 1:
                await asyncio.sleep(retry_delay)
                
        if missing:
            msg = f"Missing critical tables after retries: {', '.join(missing)}. Please apply migrations."
            logger.error(msg)
            raise RuntimeError(msg)
            
    except Exception as e:
        logger.error("Database schema verification critically failed: %s", e)
        raise

    # Verify migration version matches code
    try:
        from alembic import config
        from alembic.migration import MigrationContext
        
        # Assume alembic.ini is in the root or accessible
        alembic_cfg = config.Config("alembic.ini") 
        # We need to set the script location relative to where we are or use absolute path
        # If alembic.ini relies on recursive search or specific paths, this might be tricky.
        # Simplification: Just verify alembic_version table is readable and has a revision.
        with engine.connect() as connection:
             context = MigrationContext.configure(connection)
             current_rev = context.get_current_revision()
             logger.info(f"Current database revision: {current_rev}")
             if not current_rev:
                 logger.warning("Database has no migration revision! Changes may be missing.")
    except Exception as e:
        logger.warning(f"Could not verify migration version: {e}")
        
    # Patchset 107: Secondary startup validation for production safety
    if settings.APP_ENV in ("production", "staging"):
        if getattr(settings, "USE_MOCK", False):
            raise RuntimeError("FATAL: Refusing startup in production/staging with USE_MOCK=True.")
        if not getattr(settings, "REQUIRE_REAL_LLM", False):
            raise RuntimeError("FATAL: Refusing startup in production/staging with REQUIRE_REAL_LLM=False.")
        if not getattr(settings, "ENABLE_SEC_HEADERS", True):
            raise RuntimeError("FATAL: Refusing startup in production/staging with ENABLE_SEC_HEADERS=False.")
        if not getattr(settings, "ENABLE_CSRF", True):
            logger.warning("WARNING: Startup permitted with ENABLE_CSRF=False but this is not recommended in production.")
            
    _warn_on_multi_worker()
    try:
        ensure_rate_limiter_ready(raise_on_failure=settings.RATE_LIMIT_BACKEND == "redis")
    except Exception as exc:
        logger.error("Rate limiter backend check failed: %s", exc)
        if not settings.IS_LOCAL_ENV:
            raise
    try:
        models = list_enabled_models()
        if not models:
            logger.error("No models enabled; configure provider API keys.")
        else:
            logger.info("Models enabled: %s (default=%s)", [m.id for m in models], get_default_model().id)
    except Exception as exc:
        logger.error("Model registry initialization failed: %s", exc)
    
    # Initialize and start SSE backend (singleton)
    sse_backend = get_sse_backend()
    await sse_backend.start()
    app.state.sse_backend = sse_backend
    
    # Patchset 66.0: Start stale debate cleanup loop
    from orchestrator_cleanup import start_cleanup_loop, stop_cleanup_loop
    cleanup_task = start_cleanup_loop()
    
    try:
        yield
    finally:
        # Patchset 66.0: Stop cleanup loop gracefully
        stop_cleanup_loop()
        if cleanup_task and not cleanup_task.done():
            with suppress(asyncio.CancelledError):
                await cleanup_task
        await sse_backend.stop()

        # Centralized Redis connection pool shutdown
        try:
            from redis_pool import close_async_redis, close_sync_redis
            await close_async_redis()
            close_sync_redis()
            logger.info("Redis connection pools closed successfully during lifespan shutdown.")
        except Exception as exc:
            logger.warning("Error closing Redis connection pools during lifespan shutdown: %s", exc)
app = FastAPI(
    title="Consultaion API",
    version="0.1.0",
    lifespan=lifespan,
    dependencies=[] if TEST_FAST_APP else [Depends(csrf_protect)],
)

if ENABLE_SEC_HEADERS and not TEST_FAST_APP:

    @app.middleware("http")
    async def security_headers_middleware(request: Request, call_next):
        response: Response = await call_next(request)
        response.headers.setdefault("X-Content-Type-Options", "nosniff")
        response.headers.setdefault("X-Frame-Options", "DENY")
        response.headers.setdefault("Referrer-Policy", "strict-origin-when-cross-origin")
        response.headers.setdefault("X-XSS-Protection", "0")
        return response


def _warn_on_multi_worker() -> None:
    """Warn if multiple workers are in use while SSE queues are process-local."""
    worker_env = settings.WEB_CONCURRENCY or settings.GUNICORN_WORKERS
    try:
        if worker_env and int(worker_env) > 1 and settings.SSE_BACKEND.lower() == "memory":
            logger.warning(
                "SSE_BACKEND=memory; running with %s workers may cause /debates/{id}/stream to miss events. "
                "Switch to SSE_BACKEND=redis for multi-worker deployments.",
                worker_env,
            )
    except ValueError:
        return


class RequestIDMiddleware(BaseHTTPMiddleware):
    async def dispatch(self, request: Request, call_next):
        clear_log_context()
        request_id = request.headers.get("x-request-id") or str(uuid.uuid4())
        token = set_request_id(request_id)
        try:
            response = await call_next(request)
        finally:
            reset_request_id(token)
            clear_log_context()
        response.headers["X-Request-ID"] = request_id
        return response


origins = settings.CORS_ORIGINS.split(",")
app.add_middleware(RequestIDMiddleware)
if settings.ENV != "test":
    app.add_middleware(BodySizeLimitMiddleware, max_content_size=10 * 1024 * 1024) # 10MB
    app.add_middleware(
        CORSMiddleware,
        allow_origins=origins,
        allow_credentials=True,
        allow_methods=["*"],
        allow_headers=["*"],
    )

from exceptions import AppError

from exception_handlers import app_error_handler, http_exception_handler, validation_exception_handler
from fastapi.exceptions import RequestValidationError
from starlette.exceptions import HTTPException

app.add_exception_handler(AppError, app_error_handler)
app.add_exception_handler(HTTPException, http_exception_handler)
app.add_exception_handler(RequestValidationError, validation_exception_handler)

# Domain routers live in apps/api/routes/*
from routes.debug import router as debug_router  # Patchset 53.0

app.include_router(auth_router)
app.include_router(ops_router)
app.include_router(stats_router)
app.include_router(models_router)
app.include_router(debates_router)
app.include_router(teams_router)
app.include_router(admin_router)

# Patchset 53.0: Debug routes (only registered in safe environments)
if settings.IS_LOCAL_ENV or settings.AUTH_DEBUG:
    app.include_router(debug_router)

# Import and add routing admin router
from routes.routing_admin import router as routing_admin_router

app.include_router(routing_admin_router)

app.include_router(billing_router)
app.include_router(promotions_router)
app.include_router(api_keys_router)

from routes.features import router as features_router
from routes.gifs import router as gifs_router

app.include_router(gifs_router, prefix="/gifs", tags=["gifs"])
app.include_router(features_router)

# Patchset 77: Conversation V2 voting API
from routes.votes import router as votes_router
app.include_router(votes_router)

# Phase 4: API Versioning Strategy - Mount API v1 namespace router
from fastapi import APIRouter

v1_router = APIRouter(prefix="/api/v1")
v1_router.include_router(auth_router)
v1_router.include_router(ops_router)
v1_router.include_router(stats_router)
v1_router.include_router(models_router)
v1_router.include_router(debates_router)
v1_router.include_router(teams_router)
v1_router.include_router(admin_router)
if settings.IS_LOCAL_ENV or settings.AUTH_DEBUG:
    v1_router.include_router(debug_router)
v1_router.include_router(routing_admin_router)
v1_router.include_router(billing_router)
v1_router.include_router(promotions_router)
v1_router.include_router(api_keys_router)
v1_router.include_router(gifs_router, prefix="/gifs", tags=["gifs"])
v1_router.include_router(features_router)
v1_router.include_router(votes_router)

app.include_router(v1_router)




# Lifespan helpers


__all__ = [
    # routers (for compatibility/imports)
    "auth_router",
    "stats_router",
    "models_router",
    "debates_router",
    "teams_router",
    "admin_router",
    "billing_router",
    "promotions_router",
    # exported handlers for legacy imports/tests
    "AuthRequest",
    "get_me",
    "login_user",
    "logout_user",
    "register_user",
    "google_login",
    "google_callback",
    "get_model_leaderboard_stats",
    "get_model_detail",
    "get_debate_summary",
    "get_rate_limit_snapshot",
    "get_system_health",
    "get_hall_of_fame_stats",
    "get_debate_report",
    "export_debate_report",
    "get_debate_events",
    "get_debate_judges",
    "export_scores_csv",
    "stream_events",
    "create_debate",
    "DebateCreate",
    "DebateUpdate",
    "list_debates",
    "get_debate",
    "start_debate_run",
    "get_leaderboard",
    "get_leaderboard_persona",
    "update_debate",
    "TeamCreate",
    "TeamMemberCreate",
    "create_team",
    "list_teams",
    "list_team_members",
    "add_team_member",
    "admin_users",
    "admin_logs",
    "admin_ops_summary",
    "update_ratings_endpoint",
]